evil ass rewrite of the recipebook no more 3000 line file.
