+++ tgui/packages/tgui/interfaces/RecipeBook/shared.ts
@@ -1,0 +1,10 @@
+import type { Recipe } from './types';
+
+export type NavProps = {
+  r: Recipe;
+  lookup: Map<string, Recipe>;
+  pickerMap: Map<string, Recipe[]>;
+  allRecipes: Recipe[];
+  essenceIndex: Map<string, Recipe[]>;
+  nav: (r: Recipe) => void;
+};
